Add modern-es6-only config to be used by Vue 3 apps
Vue 3 in particular _requires_ the Proxy object, so trying to
transpile/polyfill for browsers that do not support Proxy is pointless
as the Proxy object cannot be polyfilled/transpiled.

This config is probably somewhat temporary until T178356 is implemented.
That is: Raising Grad A JavaScript requirement from ES5 to ES6 (ES2015).

Further, it seems that caniuse is only tracking the most recent Chrome
version for its "Android" (and "Chrome for Android") column. This makes
it a bit hard to provide a reliable value there. For now, this patch
uses the same approach for Android as for Chrome itself. However, we
should reconsider whether the "last 2 versions" approach is the right
thing to do in general, given the fast release cycles of these browsers.

See also:
https://caniuse.com/es6
https://caniuse.com/proxy